Update ui package to use typescript
