+++ packages/ui/src/Footer.tsx
@@ -1,0 +1,187 @@
+import React from "react";
+import { Grid, Typography } from "@material-ui/core";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { makeStyles } from "@material-ui/core/styles";
+
+import { Link } from "./Link";
+import { EmailLink } from "./EmailLink";
+
+import PrivateWrapper from "../../../apps/platform/src/components/PrivateWrapper";
+
+const useStyles = makeStyles((theme) => ({
+  footer: {
+    // @ts-ignore
+    backgroundColor: theme.palette.footer,
+    color: "#fff",
+    margin: 0,
+    width: "100%",
+    padding: 24,
+  },
+}));
+
+const useLinkStyles = makeStyles(() => ({
+  iconClass: {
+    marginRight: "10px",
+  },
+  linkContainer: {
+    marginBottom: "8px",
+  },
+}));
+
+let FooterLink = ({ label, url, icon }) => {
+  const classes = useLinkStyles();
+  return (
+    <Grid item xs={12} className={classes.linkContainer}>
+      <Typography color="inherit">
+        {url.startsWith("mailto") ? (
+          <EmailLink href={url} label={label} icon={icon} />
+        ) : (
+          <Link external footer to={url}>
+            {icon && (
+              <FontAwesomeIcon
+                className={classes.iconClass}
+                icon={icon}
+                size="lg"
+              />
+            )}
+            {label}
+          </Link>
+        )}
+      </Typography>
+    </Grid>
+  );
+};
+
+const FooterSectionHeading = ({ children }) => (
+  <Grid item xs={12}>
+    <Typography variant="h6" color="inherit">
+      {children}
+    </Typography>
+  </Grid>
+);
+
+const useSocialLinkStyle = makeStyles(() => ({
+  iconsContainer: {
+    maxWidth: "235px",
+  },
+  socialIcon: {
+    fontSize: "30px",
+    color: "white",
+  },
+}));
+
+let FooterSocial = ({ social }) => {
+  const classes = useSocialLinkStyle();
+  return (
+    <>
+      <FooterSectionHeading>Follow us</FooterSectionHeading>
+      <Grid
+        className={classes.iconsContainer}
+        container
+        justifyContent="space-between"
+      >
+        {social.map(({ icon, url }, i) => (
+          <Grid item key={i}>
+            <Link external footer to={url}>
+              <FontAwesomeIcon className={classes.socialIcon} icon={icon} />
+            </Link>
+          </Grid>
+        ))}
+      </Grid>
+    </>
+  );
+};
+
+const useSectionStyles = makeStyles({
+  section: {
+    width: "100%",
+  },
+});
+
+const FooterSection = ({
+  heading,
+  links,
+  social,
+}: {
+  heading: React.ReactNode;
+  links: {
+    showOnlyPartner: boolean;
+    label: string;
+    icon: unknown;
+    url: string;
+  }[];
+  social?: unknown;
+}) => {
+  const classes = useSectionStyles();
+  return (
+    <Grid
+      item
+      xs={12}
+      sm={6}
+      md={3}
+      container
+      direction="column"
+      justifyContent="space-between"
+    >
+      <Grid item className={classes.section}>
+        <FooterSectionHeading>{heading}</FooterSectionHeading>
+        {links.map((link, i) => {
+          if (link.showOnlyPartner) {
+            return (
+              <PrivateWrapper key={i}>
+                <FooterLink
+                  label={link.label}
+                  url={link.url}
+                  icon={link.icon}
+                />
+              </PrivateWrapper>
+            );
+          } else {
+            return (
+              <FooterLink
+                key={i}
+                label={link.label}
+                url={link.url}
+                icon={link.icon}
+              />
+            );
+          }
+        })}
+      </Grid>
+
+      {social ? (
+        <Grid item>
+          <FooterSocial social={social} />
+        </Grid>
+      ) : null}
+    </Grid>
+  );
+};
+
+const Footer = ({ externalLinks }) => {
+  const classes = useStyles();
+  return (
+    <Grid
+      className={classes.footer}
+      container
+      justifyContent="center"
+      spacing={3}
+    >
+      <Grid item container xs={12} md={10} spacing={2}>
+        <FooterSection heading="About" links={externalLinks.about} />
+        <FooterSection
+          heading="Help"
+          links={externalLinks.help}
+          social={externalLinks.social}
+        />
+        <FooterSection heading="Partners" links={externalLinks.partners} />
+        <FooterSection
+          heading="About Open Targets"
+          links={externalLinks.network}
+        />
+      </Grid>
+    </Grid>
+  );
+};
+
+export default Footer;
